% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast.R
\name{cast}
\alias{cast}
\title{Search for Cast Memebers}
\usage{
cast(id = "1")
}
\arguments{
\item{id}{The id of the character (portrayed by the actor/actress the user is interested in) the user wants to search for, it can also be the full name of the character.}
}
\value{
A table that shows some detailed information about an actor or actress who portrayed the character the user gave input to the function.
}
\description{
If the user is particularly interested in one of the actor or actresss appeared in the show, he or she can use this function to search for some detailed information about that actor or actress.
}
\examples{
cast()
cast("10")
cast("38")
cast("Jesse Pinkman")
}
\author{
Chuze Zhang <\url{https://github.com/QMSS-G5072-2019/Zhang_Chuze}>
}
