% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/death.R
\name{cause}
\alias{cause}
\title{Helper Function Search Death Scenarios by the Cause}
\usage{
cause(df, weapon)
}
\arguments{
\item{df}{A data table that contains all the death scenarios}

\item{weapon}{Specify the cause of death, select from gun, knive, bomb, poison, strangle, crush, beat, decapitate, overdose, slit and collide}
}
\value{
All death senarios caused by *weapon*
}
\description{
This function is a helper function that helps the main function to search for death scenarios by a specific cause
}
\author{
Chuze Zhang <\url{https://github.com/QMSS-G5072-2019/Zhang_Chuze}>
}
\keyword{internal}
