% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characters.R
\name{character}
\alias{character}
\title{Information Search}
\usage{
character(id = "1", func = "display")
}
\arguments{
\item{id}{The id of the character the user wants to search for, it can also be the full name of the character. (like 'Walter White')}

\item{func}{Specify what exactly the user wants to know about this character, valid inputs include display(all available information), occupation, status, death count, portrayed and image. Display is the default.}
}
\value{
A table includes the requested(specified by *func*) information about a character.
}
\description{
This function allows the user to search for information they want to know about a character.
}
\examples{
character()
character(5, "display")
character(7, "occupation")
character(9, "image")
}
\author{
Chuze Zhang <\url{https://github.com/QMSS-G5072-2019/Zhang_Chuze}>
}
