% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast.R
\name{child_cleanup}
\alias{child_cleanup}
\title{Helper Function Clean Up}
\usage{
child_cleanup(value)
}
\arguments{
\item{value}{Digits representing how many children an individual has, or just the name of a child in character string}
}
\value{
Keep the digits and remove all explanations after that.
}
\description{
This function is a helper function that helps the main function to clean up the children section
}
\author{
Chuze Zhang <\url{https://github.com/QMSS-G5072-2019/Zhang_Chuze}>
}
\keyword{internal}
