% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characters.R
\name{convert_url}
\alias{convert_url}
\title{Helper Function Convert}
\usage{
convert_url(na, endp)
}
\arguments{
\item{na}{character input, the name of the character}

\item{endp}{specify the API endpoint of the search}
}
\value{
content of the search in list form
}
\description{
This function is a helper function that helps the main function to convert character inputs to readable api addresses.
}
\author{
Chuze Zhang <\url{https://github.com/QMSS-G5072-2019/Zhang_Chuze}>
}
\keyword{internal}
