% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/death.R
\name{death}
\alias{death}
\title{All Death Scenarios}
\usage{
death(func = "random", weapon = "gun")
}
\arguments{
\item{func}{Specify what the user wants this function to perform, valid inputs include random(some information about a death at random), all(retrieve information about every death happened in Breaking Bad) and search(search for death scenarios by a specific cause)}

\item{weapon}{The casue of death, valid input include ("gun","shot","stab","knive","bomb","poison","strangle","garrote","crush","beat","decapitate","overdose","slit","collide")}
}
\value{
A table includes the requested(specified by *func*) information.
}
\description{
This function allows the user to retrieve information about some death scenarios that happened in the Breaking Bad
}
\examples{
death()
death("random")
death("all")
death("search", "bomb")
}
\author{
Chuze Zhang <\url{https://github.com/QMSS-G5072-2019/Zhang_Chuze}>
}
