% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characters.R
\name{option}
\alias{option}
\title{Full Character List}
\usage{
option(limit = 63)
}
\arguments{
\item{limit}{limit the amount of characters you receive. default to the full list (63).}
}
\value{
A table includes some basic information about the characters that appeared in the show.
}
\description{
This function allows the user to see the full character list. The output table will include some basic information about the characters that appeared in the show.
}
\examples{
option()
option("10")
}
\author{
Chuze Zhang <\url{https://github.com/QMSS-G5072-2019/Zhang_Chuze}>
}
